ajustando rotas e controller

+++ app/Http/Controllers/ProjectController.php
@@ -37,7 +37,8 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
+
+        return $this->repository->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
     }
 
     /**
@@ -69,7 +70,9 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
-
+        if($this->checkProjectOwner($id)==false){
+            return ['error' => 'Access Forbidden'];
+        }
         return $this->repository->find($id);
     }
 
@@ -93,6 +96,9 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if($this->checkProjectOwner($id)==false){
+            return ['error' => 'Access Forbidden'];
+        }
         return $this->clientService->update($request->all(), $id);
     }
 
@@ -104,6 +110,16 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if($this->checkProjectOwner($id)==false){
+            return ['error' => 'Access Forbidden'];
+        }
         return $this->repository->delete($id);
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($projectId, $userId);
     }
 }
